requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
drf-spectacular>=0.27,<0.28
django-environ>=0.11,<0.12
django-htmx>=1.18,<1.19
pillow>=10.4,<10.5

PyMySQL>=1.1,<1.2

pytest>=8.3,<8.4
pytest-django>=4.8,<4.9
factory-boy>=3.3,<3.4

black>=24.8,<24.9
isort>=5.13,<5.14
flake8>=7.1,<7.2
mypy>=1.11,<1.12
